[Unified Observability] Clean overview page feature toggle (#130650)
* Move old overview page to index

* remove overview page feature toggle

* Remove code for old alerts section

* Fix translations

* remove feature toggle from tests
