+++ x-pack/plugins/observability/public/application/application.test.tsx
@@ -63,7 +63,6 @@ describe('renderApp', () => {
       unsafe: {
         alertingExperience: { enabled: true },
         cases: { enabled: true },
-        overviewNext: { enabled: false },
         rules: { enabled: true },
       },
     };
