+++ x-pack/plugins/observability/public/components/app/section/apm/index.test.tsx
@@ -48,7 +48,6 @@ describe('APMSection', () => {
         unsafe: {
           alertingExperience: { enabled: true },
           cases: { enabled: true },
-          overviewNext: { enabled: false },
           rules: { enabled: true },
         },
       },
