+++ x-pack/plugins/observability/public/pages/overview/overview.stories.tsx
@@ -85,7 +85,6 @@ const withCore = makeDecorator({
                 unsafe: {
                   alertingExperience: { enabled: true },
                   cases: { enabled: true },
-                  overviewNext: { enabled: false },
                   rules: { enabled: true },
                 },
               },
