+++ x-pack/plugins/observability/public/utils/test_helper.tsx
@@ -26,7 +26,6 @@ const config = {
   unsafe: {
     alertingExperience: { enabled: true },
     cases: { enabled: true },
-    overviewNext: { enabled: false },
     rules: { enabled: true },
   },
 };
