+++ x-pack/plugins/translations/translations/fr-FR.json
@@ -21658,12 +21658,6 @@
     "xpack.observability.noDataConfig.beatsCard.title": "Ajouter des intégrations",
     "xpack.observability.noDataConfig.solutionName": "Observabilité",
     "xpack.observability.notAvailable": "S. O.",
-    "xpack.observability.overview.alert.allTypes": "Tous les types",
-    "xpack.observability.overview.alert.appLink": "Afficher toutes les alertes",
-    "xpack.observability.overview.alert.errorMessage": "Une erreur s’est produite lors du chargement des données d’alerte. Réessayez plus tard.",
-    "xpack.observability.overview.alert.errorTitle": "Impossible de charger les données d’alerte.",
-    "xpack.observability.overview.alerts.appLink": "Afficher les alertes",
-    "xpack.observability.overview.alerts.muted": "Muet",
     "xpack.observability.overview.alerts.title": "Alertes",
     "xpack.observability.overview.apm.appLink": "Afficher l’inventaire des services",
     "xpack.observability.overview.apm.services": "Services",
