+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -21815,12 +21815,6 @@
     "xpack.observability.noDataConfig.beatsCard.title": "統合の追加",
     "xpack.observability.noDataConfig.solutionName": "Observability",
     "xpack.observability.notAvailable": "N/A",
-    "xpack.observability.overview.alert.allTypes": "すべてのタイプ",
-    "xpack.observability.overview.alert.appLink": "すべてのアラートを表示",
-    "xpack.observability.overview.alert.errorMessage": "アラートデータの読み込みエラーが発生しました。しばらくたってから再試行してください。",
-    "xpack.observability.overview.alert.errorTitle": "アラートデータを読み込めませんでした",
-    "xpack.observability.overview.alerts.appLink": "アラートを表示",
-    "xpack.observability.overview.alerts.muted": "ミュート",
     "xpack.observability.overview.alerts.title": "アラート",
     "xpack.observability.overview.apm.appLink": "サービスインベントリを表示",
     "xpack.observability.overview.apm.services": "サービス",
