+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -21847,12 +21847,6 @@
     "xpack.observability.noDataConfig.beatsCard.title": "添加集成",
     "xpack.observability.noDataConfig.solutionName": "Observability",
     "xpack.observability.notAvailable": "不可用",
-    "xpack.observability.overview.alert.allTypes": "所有类型",
-    "xpack.observability.overview.alert.appLink": "显示所有告警",
-    "xpack.observability.overview.alert.errorMessage": "加载告警数据时出现错误。请稍后重试。",
-    "xpack.observability.overview.alert.errorTitle": "我们无法加载告警数据",
-    "xpack.observability.overview.alerts.appLink": "显示告警",
-    "xpack.observability.overview.alerts.muted": "已静音",
     "xpack.observability.overview.alerts.title": "告警",
     "xpack.observability.overview.apm.appLink": "显示服务库存",
     "xpack.observability.overview.apm.services": "服务",
